flattened out the conditions
