+++ src/main/kotlin/us/timinc/mc/cobblemon/droploottables/lootconditions/PokemonPropertiesLootCondition.kt
@@ -1,0 +1,36 @@
+package us.timinc.mc.cobblemon.droploottables.lootconditions
+
+import com.cobblemon.mod.common.api.pokemon.PokemonProperties
+import com.cobblemon.mod.common.pokemon.Pokemon
+import com.cobblemon.mod.common.util.toProperties
+import com.mojang.serialization.Codec
+import com.mojang.serialization.MapCodec
+import com.mojang.serialization.codecs.RecordCodecBuilder
+import net.minecraft.loot.condition.LootCondition
+import net.minecraft.loot.condition.LootConditionType
+import net.minecraft.loot.context.LootContext
+
+class PokemonPropertiesLootCondition(
+    val properties: PokemonProperties,
+) : LootCondition {
+    companion object {
+        object KEYS {
+            const val PROPERTIES = "properties"
+        }
+
+        val CODEC: MapCodec<PokemonPropertiesLootCondition> = RecordCodecBuilder.mapCodec { instance ->
+            instance.group(
+                Codec.STRING.fieldOf(KEYS.PROPERTIES).forGetter { it.properties.originalString }
+            ).apply(instance) { PokemonPropertiesLootCondition(it.toProperties()) }
+        }
+    }
+
+    override fun test(context: LootContext): Boolean {
+        val pokemon: Pokemon = context.get(LootConditions.PARAMS.SLAIN_POKEMON)!!
+        return properties.matches(pokemon)
+    }
+
+    override fun getType(): LootConditionType {
+        return LootConditions.POKEMON_PROPERTIES
+    }
+}
